wip: timestamp adapt MemTable, Write-Ahead Log, and Read Path

+++ src/main/scala/io/github/leibnizhu/tinylsm/utils/ByteTransOps.scala
@@ -50,15 +50,15 @@ object ByteTransOps {
     (safeB3 << 24) + (safeB2 << 16) + (safeB1 << 8) + safeB0
   }
 
-  def bytesToLong(b7: Byte, b6: Byte, b5: Byte, b4: Byte, b3: Byte, b2: Byte, b1: Byte, b0: Byte): Int = {
-    val safeB7: Int = if (b7 < 0) b7 + 256 else b7
-    val safeB6: Int = if (b6 < 0) b6 + 256 else b6
-    val safeB5: Int = if (b5 < 0) b5 + 256 else b5
-    val safeB4: Int = if (b4 < 0) b4 + 256 else b4
-    val safeB3: Int = if (b3 < 0) b3 + 256 else b3
-    val safeB2: Int = if (b2 < 0) b2 + 256 else b2
-    val safeB1: Int = if (b1 < 0) b1 + 256 else b1
-    val safeB0: Int = if (b0 < 0) b0 + 256 else b0
+  def bytesToLong(b7: Byte, b6: Byte, b5: Byte, b4: Byte, b3: Byte, b2: Byte, b1: Byte, b0: Byte): Long = {
+    val safeB7: Long = if (b7 < 0) b7 + 256 else b7
+    val safeB6: Long = if (b6 < 0) b6 + 256 else b6
+    val safeB5: Long = if (b5 < 0) b5 + 256 else b5
+    val safeB4: Long = if (b4 < 0) b4 + 256 else b4
+    val safeB3: Long = if (b3 < 0) b3 + 256 else b3
+    val safeB2: Long = if (b2 < 0) b2 + 256 else b2
+    val safeB1: Long = if (b1 < 0) b1 + 256 else b1
+    val safeB0: Long = if (b0 < 0) b0 + 256 else b0
     (safeB7 << 56) + (safeB6 << 48) + (safeB5 << 40) + (safeB4 << 32) + (safeB3 << 24) + (safeB2 << 16) + (safeB1 << 8) + safeB0
   }
 }
